Detail screen Date widget optimisation
