+++ app/lib/features/events/widgets/event_item.dart
@@ -1,8 +1,8 @@
 import 'package:acter/common/providers/room_providers.dart';
 import 'package:acter/common/utils/routes.dart';
 import 'package:acter/common/utils/utils.dart';
-import 'package:acter/features/events/event_utils/event_utils.dart';
 import 'package:acter/features/events/providers/event_providers.dart';
+import 'package:acter/features/events/widgets/event_date_widget.dart';
 import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart'
     show CalendarEvent;
 import 'package:flutter/material.dart';
@@ -44,7 +44,7 @@ class EventItem extends StatelessWidget {
         child: Row(
           mainAxisAlignment: MainAxisAlignment.start,
           children: [
-            _buildEventDate(context),
+            EventDateWidget(calendarEvent: event),
             Expanded(
               child: Column(
                 mainAxisAlignment: MainAxisAlignment.center,
@@ -58,29 +58,6 @@ class EventItem extends StatelessWidget {
             const SizedBox(width: 10),
             if (isShowRsvp) _buildRsvpStatus(context),
             const SizedBox(width: 10),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildEventDate(BuildContext context) {
-    final day = getDayFromDate(event.utcStart());
-    final month = getMonthFromDate(event.utcStart());
-
-    return Card(
-      margin: const EdgeInsets.all(12),
-      color: getColorBasedOnType(context),
-      child: Container(
-        height: 70,
-        width: 70,
-        alignment: Alignment.center,
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: [
-            Text(month),
-            Text(day, style: Theme.of(context).textTheme.titleLarge),
           ],
         ),
       ),
@@ -149,17 +126,5 @@ class EventItem extends StatelessWidget {
       }
     }
     return L10n.of(context).pending;
-  }
-
-  Color getColorBasedOnType(BuildContext context) {
-    if (getEventType(event) == EventFilters.ongoing) {
-      return Colors.green;
-    } else if (getEventType(event) == EventFilters.upcoming) {
-      return Colors.yellow.shade800;
-    } else if (getEventType(event) == EventFilters.past) {
-      return Colors.grey.shade600;
-    } else {
-      return Theme.of(context).primaryColor;
-    }
   }
 }
